create polars equivalent of pandas_flavor
