+++ janitor/polars/__init__.py
@@ -1,12 +1,13 @@
-from .dataframe import PolarsDataFrame
-from .expressions import PolarsExpr
-from .lazyframe import PolarsLazyFrame
-from .pivot_longer import pivot_longer_spec
+from .clean_names import clean_names, make_clean_names
+from .complete import complete
+from .pivot_longer import pivot_longer, pivot_longer_spec
+from .row_to_names import row_to_names
 
 __all__ = [
     "pivot_longer_spec",
+    "pivot_longer",
     "clean_names",
-    "PolarsDataFrame",
-    "PolarsLazyFrame",
-    "PolarsExpr",
+    "make_clean_names",
+    "row_to_names",
+    "complete",
 ]
